feat: Add default value for booking_time column in Booking model
Set the default value of the booking_time column in the Booking model to the current time using the `datetime.utcnow()` function. This ensures that the booking_time is automatically set to the current time when a new booking is created.

Also, update the create_booking function in the booking_service module to set the booking_time to the current time when creating a new booking.

This change improves the accuracy of the booking time and simplifies the code by removing the need to manually set the booking_time.

Fixes #123